+++ Extension/src/LanguageServer/referencesTreeDataProvider.ts
@@ -4,48 +4,12 @@
  * ------------------------------------------------------------------------------------------ */
 'use strict';
 import * as vscode from 'vscode';
-import * as util from '../common';
 import { ReferencesModel, TreeNode, NodeType } from './referencesModel';
-import { ReferenceType, getReferenceTagString } from './references';
+import { ReferenceType, getReferenceTagString, getReferenceItemIconPath } from './references';
 import * as nls from 'vscode-nls';
 
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
 const localize: nls.LocalizeFunc = nls.loadMessageBundle();
-
-export function getReferenceTypeIconPath(referenceType: ReferenceType): { light: string; dark: string } {
-    const assetsFolder: string = "assets/";
-    const postFixLight: string = "-light.svg";
-    const postFixDark: string = "-dark.svg";
-    let basePath: string = "ref-cannot-confirm";
-
-    switch (referenceType) {
-        case ReferenceType.Confirmed: basePath = "ref-confirmed"; break;
-        case ReferenceType.Comment: basePath = "ref-comment"; break;
-        case ReferenceType.String: basePath = "ref-string"; break;
-        case ReferenceType.Inactive: basePath = "ref-inactive"; break;
-        case ReferenceType.CannotConfirm: basePath = "ref-cannot-confirm"; break;
-        case ReferenceType.NotAReference: basePath = "ref-not-a-reference"; break;
-        case ReferenceType.ConfirmationInProgress: basePath = "ref-confirmation-in-progress"; break;
-    }
-
-    let lightPath: string = util.getExtensionFilePath(assetsFolder + basePath + postFixLight);
-    let darkPath: string = util.getExtensionFilePath(assetsFolder + basePath + postFixDark);
-    return {
-        light: lightPath,
-        dark: darkPath
-    };
-}
-
-function getReferenceCanceledIconPath(): { light: string; dark: string } {
-    return {
-        light: util.getExtensionFilePath("assets/ref-canceled-light.svg"),
-        dark: util.getExtensionFilePath("assets/ref-canceled-dark.svg")
-    };
-}
-
-function getReferenceItemIconPath(type: ReferenceType, isCanceled: boolean): { light: string; dark: string } {
-    return (isCanceled && type === ReferenceType.ConfirmationInProgress) ? getReferenceCanceledIconPath() : getReferenceTypeIconPath(type);
-}
 
 export class ReferencesTreeDataProvider implements vscode.TreeDataProvider<TreeNode> {
     private referencesModel: ReferencesModel | undefined;
